+++ nmon-logger-rsyslog/etc/nmon-logger/bin/nmon_external_cmd/nmon_external_snap.sh
@@ -9,10 +9,8 @@
 
 # 2017/04/29, Guilhem Marchand:         - AIX compatibility issues, detach the commands in background
 # 2017/06/04, Guilhem Marchand:         - Manage nmon external data in a dedicated file
-# 2017/06/05, V1.0.3: Guilhem Marchand:
-#                                          - Mirror update of the TA-nmon
 
-# Version 1.0.3
+# Version 1.0.2
 
 # For AIX / Linux / Solaris
 
@@ -37,3 +35,23 @@ echo "PROCCOUNT,$1,`ps -ef | wc -l`" >>NMON_FIFO_PATH/nmon_external.dat &
 
 # Uptime information (uptime command output)
 echo "UPTIME,$1,\"`uptime | sed 's/^\s//g' | sed 's/,/;/g'`\"" >>NMON_FIFO_PATH/nmon_external.dat &
+
+# df table information
+DF_TABLE=`df -k -P | sed '1d' | egrep -v '\/proc$|/dev$|\/run$|^tmpfs.*\/dev.*$|^tmpfs.*\/run.*$|^tmpfs.*\/sys.*$|^tmpfs.*\/var.*$' | awk '{print $6}'`
+for fs in $DF_TABLE; do
+    echo "DF_STORAGE,$1,`df -k -P $fs | sed '1d' | sed 's/%//g' | sed 's/,/;/g' | awk '{print $1 "," $2 "," $3 "," $4 "," $5 "," $6}'`" >>NMON_FIFO_PATH/nmon_external.dat
+done
+
+# DF_INODES, for AIX and Linux
+case `uname` in
+"AIX")
+    for fs in $DF_TABLE; do
+        echo "DF_INODES,$1,`df -i $fs | sed '1d' | sed 's/%//g' | sed 's/,/;/g' | awk '{print $1 "," $5 "," $6 "," $7}'`" >>NMON_FIFO_PATH/nmon_external.dat
+    done
+    ;;
+"Linux")
+    for fs in $DF_TABLE; do
+        echo "DF_INODES,$1,`df -i -P $fs | sed '1d' | sed 's/%//g' | sed 's/,/;/g' | awk '{print $1 "," $2 "," $3 "," $4 "," $5 "," $6}'`" >>NMON_FIFO_PATH/nmon_external.dat
+    done
+    ;;
+esac