+++ nmon-logger-splunk-hec/etc/nmon-logger/bin/nmon2kv.py
@@ -43,7 +43,11 @@
 # - 2017/27/07: V1.0.7: Guilhem Marchand:
 #                                           - Splunk HEC implementation
 # - 2017/10/08: V1.0.8: Guilhem Marchand:
-#                                           - Fix epoch timestmap failure for dynamic sections
+#                                           - Fix epoch timestamp failure for dynamic sections
+# - 2017/12/08: V1.0.9: Guilhem Marchand:
+#                                           - Preserve data order during key value transformation
+# - 08/19/2017: V1.0.10: Guilhem Marchand:
+#                                           - fix: improve header check for static sections
 
 # Load libs
 
@@ -64,7 +68,7 @@ import json
 import subprocess
 
 # Converter version
-nmon2kv_version = '1.0.8'
+nmon2kv_version = '1.0.10'
 
 # LOGGING INFORMATION:
 # - The program uses the standard logging Python module to display important messages in Splunk logs
@@ -464,31 +468,30 @@ def write_kv(input, kv_file):
 
         if kvdelim:
             for row in reader:
+                data = ""
                 for k, v in row.items():
-                    f.write("%s=\"%s\" " % (k, v))
-                f.write('\n')
+                    data = ("%s=\"%s\" " % (k, v)) + data
+                f.write(data + '\n')
         else:
             for row in reader:
+                data = ""
                 for k, v in row.items():
-                    f.write("%s=%s " % (k, v))
-                f.write('\n')
+                    data = ("%s=%s " % (k, v)) + data
+                f.write(data + '\n')
 
 # Stream to Splunk HEC
 def write_kv_to_http(input):
 
     reader = csv.DictReader(input)
-    final_data = ""
     http_data = ""
 
     for row in reader:
-
+        data = ""
         for k, v in row.items():
-
-            data="%s=\"%s\" " % (k, v)
-            final_data = final_data + data
+            data = ("%s=\"%s\" " % (k, v)) + data
 
         # extract epochtime
-        timestamp_match = re.match(r'.*timestamp="([0-9]*)".*', final_data)
+        timestamp_match = re.match(r'.*timestamp="([0-9]*)".*', data)
         if timestamp_match:
             timestamp = timestamp_match.group(1)
         else:
@@ -496,14 +499,11 @@ def write_kv_to_http(input):
             timestamp = time.strftime("%s")
 
         # escape any double quote
-        params = final_data.replace('"', '\\"')
+        params = data.replace('"', '\\"')
 
         # This might be changed for a more Pythonic approach in the future!
         http_data = http_data + "\n" + '{"time": "' +\
                     str(timestamp) + '", "sourcetype": "nmon_data:fromhttp", "event": "' + params + '"}'
-
-        # empty variable before next iteration
-        final_data = ""
 
     with open(SPLUNK_HEC_BATCHFILE, "ab") as f:
         f.write(http_data)
@@ -1260,6 +1260,8 @@ def standard_section_fn(section):
     # Set output file
     currsection_output = NMON_VAR + '/nmon_perfdata.log'
 
+    header_found = False
+
     # Store last epochtime if in real time mode
     keyref = HOSTNAME_VAR + '/' + HOSTNAME + '.' + section + '_lastepoch.txt'
 
@@ -1383,6 +1385,9 @@ def standard_section_fn(section):
                     if header_match:
                         header = header_match.group(2)
 
+                        # header has been found
+                        header_found = True
+
                         # increment
                         count += 1
 
@@ -1400,7 +1405,7 @@ def standard_section_fn(section):
                 # Old Nmon version sometimes incorporates a Txxxx reference in the header, this is unclean
                 # but we want to try getting the header anyway
 
-                elif not fullheader_match:
+                if not header_found:
                     # Assume the header may start with Txxx, then 1 non alpha char
                     myregex = '(' + section + ')\,(T\d+),([a-zA-Z]+.+)'
                     fullheader_match = re.search(myregex, line)